Adding :always_show_create option
This shows the 'Write your own review button' regardless of if you are logged in or not. The goal
is to incentivize the user to write a review by clicking on the button and then logging into their account.
It also lets people know that they *can* write reviews (if they aren't logged in they wouldn't know they could!)